Add IPTables Rule Chain to allow custom polocies
Cattle then only manges the creation of the chain and leaves the chain
itself unmanaged, so customers are able to add their rules by themself.

This is needed for customers using lowcost hoster without having an
dedicated own network.
(so its possible for example to allow access to the complete host or certain ports for certain IPs)

fixes rancher/rancher#10387